backend-python/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
aiofiles==23.2.1
aiohttp==3.9.3
python-dotenv==1.0.1
pydantic==2.10.6
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
Pillow==11.1.0
